+++ test/libsolidity/syntaxTests/emit/emit_empty.sol
@@ -1,0 +1,7 @@
+contract C {
+  function f() public {
+    emit;
+  }
+}
+// ----
+// ParserError: (45-46): Expected event name or path.