+++ Changelog.md
@@ -1,5 +1,7 @@
 ### 0.4.25 (unreleased)
 
+Bugfixes:
+ * Type Checker: Report error when using structs in events without experimental ABIEncoderV2. This used to crash or log the wrong values.
 
 ### 0.4.24 (2018-05-16)